update post screen structure
